Determines image preview size using Retina scale

+++ piwigo/Piwigo Data Wrappers/PiwigoImageData.h
@@ -92,7 +92,7 @@ typedef enum {
 @property (nonatomic, assign) BOOL isVideo;
 
 -(NSString*)getURLFromImageSizeType:(kPiwigoImageSize)imageSize;
-+(CGFloat)widthForImageSizeType:(kPiwigoImageSize)imageSize;
++(float)widthForImageSizeType:(kPiwigoImageSize)imageSize;
 
 +(NSInteger)optimumAlbumThumbnailSizeForDevice;
 +(NSString*)nameForAlbumThumbnailSizeType:(kPiwigoImageSize)imageSize withInfo:(BOOL)addInfo;
